FIX: On multi-select paragraphs and apply a style with the bold the first sentence, only the last selected paragraph shows the bold first senence.
